requirements.txt
python = 3.10.0
django = 4.1.5
djangorestframework = 3.14.0